+++ test/reporters/code_climate_reporter_test.dart
@@ -5,6 +5,7 @@ import 'package:dart_code_metrics/src/models/code_issue.dart';
 import 'package:dart_code_metrics/src/models/code_issue_severity.dart';
 import 'package:dart_code_metrics/src/models/component_record.dart';
 import 'package:dart_code_metrics/src/models/config.dart';
+import 'package:dart_code_metrics/src/models/design_issue.dart';
 import 'package:dart_code_metrics/src/models/file_record.dart';
 import 'package:dart_code_metrics/src/models/function_record.dart';
 import 'package:dart_code_metrics/src/reporters/code_climate/code_climate_reporter.dart';
@@ -15,6 +16,8 @@ import '../stubs_builders.dart';
 
 void main() {
   group('CodeClimateReporter.report report about', () {
+    const fullPath = '/home/developer/work/project/example.dart';
+
     CodeClimateReporter _reporter;
 
     setUp(() {
@@ -23,6 +26,57 @@ void main() {
 
     test('empty file', () {
       expect(_reporter.report([]), isEmpty);
+    });
+
+    test('file with design issues', () {
+      const _issuePatternId = 'patternId1';
+      const _issuePatternDocumentation = 'https://docu.edu/patternId1.html';
+      const _issueLine = 2;
+      const _issueMessage = 'first issue message';
+      const _issueRecomendation = 'issue recomendation';
+
+      final records = [
+        FileRecord(
+          fullPath: fullPath,
+          relativePath: 'example.dart',
+          components: Map.unmodifiable(<String, ComponentRecord>{}),
+          functions: Map.unmodifiable(<String, FunctionRecord>{}),
+          issues: const [],
+          designIssue: [
+            DesignIssue(
+              patternId: _issuePatternId,
+              patternDocumentation: Uri.parse(_issuePatternDocumentation),
+              sourceSpan: SourceSpanBase(
+                  SourceLocation(1,
+                      sourceUrl: Uri.parse(fullPath),
+                      line: _issueLine,
+                      column: 3),
+                  SourceLocation(6, sourceUrl: Uri.parse(fullPath)),
+                  'issue'),
+              message: 'first issue message',
+              recommendation: _issueRecomendation,
+            ),
+          ],
+        ),
+      ];
+
+      final report =
+          (json.decode(_reporter.report(records).first) as List<Object>).first
+              as Map<String, Object>;
+
+      expect(report, containsPair('type', 'issue'));
+      expect(report, containsPair('check_name', _issuePatternId));
+      expect(report, containsPair('description', _issueMessage));
+      expect(report, containsPair('categories', ['Complexity']));
+      expect(
+          report,
+          containsPair('location', {
+            'path': 'example.dart',
+            'lines': {'begin': _issueLine, 'end': _issueLine},
+          }));
+      expect(report, containsPair('remediation_points', 50000));
+      expect(report,
+          containsPair('fingerprint', '8842a666b8aee4f2eae51205e0114dae'));
     });
 
     test('file with style severity issues', () {
@@ -34,7 +88,7 @@ void main() {
 
       final records = [
         FileRecord(
-          fullPath: '/home/developer/work/project/example.dart',
+          fullPath: fullPath,
           relativePath: 'example.dart',
           components: Map.unmodifiable(<String, ComponentRecord>{}),
           functions: Map.unmodifiable(<String, FunctionRecord>{}),
@@ -45,19 +99,17 @@ void main() {
               severity: CodeIssueSeverity.style,
               sourceSpan: SourceSpanBase(
                   SourceLocation(1,
-                      sourceUrl: Uri.parse(
-                          '/home/developer/work/project/example.dart'),
+                      sourceUrl: Uri.parse(fullPath),
                       line: _issueLine,
                       column: 3),
-                  SourceLocation(6,
-                      sourceUrl: Uri.parse(
-                          '/home/developer/work/project/example.dart')),
+                  SourceLocation(6, sourceUrl: Uri.parse(fullPath)),
                   'issue'),
               message: _issueMessage,
               correction: 'correction',
               correctionComment: 'correction comment',
             ),
           ],
+          designIssue: const [],
         ),
       ];
 
@@ -84,13 +136,15 @@ void main() {
       test('without methods', () {
         final records = [
           FileRecord(
-              fullPath: '/home/developer/work/project/example.dart',
-              relativePath: 'example.dart',
-              components: Map.unmodifiable(<String, ComponentRecord>{
-                'class': buildComponentRecordStub(methodsCount: 0),
-              }),
-              functions: Map.unmodifiable(<String, FunctionRecord>{}),
-              issues: const []),
+            fullPath: fullPath,
+            relativePath: 'example.dart',
+            components: Map.unmodifiable(<String, ComponentRecord>{
+              'class': buildComponentRecordStub(methodsCount: 0),
+            }),
+            functions: Map.unmodifiable(<String, FunctionRecord>{}),
+            issues: const [],
+            designIssue: const [],
+          ),
         ];
 
         final report =
@@ -102,13 +156,15 @@ void main() {
       test('with a lot of methods', () {
         final records = [
           FileRecord(
-              fullPath: '/home/developer/work/project/example.dart',
-              relativePath: 'example.dart',
-              components: Map.unmodifiable(<String, ComponentRecord>{
-                'class': buildComponentRecordStub(methodsCount: 20),
-              }),
-              functions: Map.unmodifiable(<String, FunctionRecord>{}),
-              issues: const []),
+            fullPath: fullPath,
+            relativePath: 'example.dart',
+            components: Map.unmodifiable(<String, ComponentRecord>{
+              'class': buildComponentRecordStub(methodsCount: 20),
+            }),
+            functions: Map.unmodifiable(<String, FunctionRecord>{}),
+            issues: const [],
+            designIssue: const [],
+          ),
         ];
 
         final report =
@@ -135,16 +191,19 @@ void main() {
     });
 
     group('function', () {
-      test('without arguments', () {
+      test('with short body', () {
         final records = [
           FileRecord(
-              fullPath: '/home/developer/work/project/example.dart',
-              relativePath: 'example.dart',
-              components: Map.unmodifiable(<String, ComponentRecord>{}),
-              functions: Map.unmodifiable(<String, FunctionRecord>{
-                'function': buildFunctionRecordStub(argumentsCount: 0),
-              }),
-              issues: const []),
+            fullPath: fullPath,
+            relativePath: 'example.dart',
+            components: Map.unmodifiable(<String, ComponentRecord>{}),
+            functions: Map.unmodifiable(<String, FunctionRecord>{
+              'function': buildFunctionRecordStub(
+                  linesWithCode: List.generate(5, (index) => index)),
+            }),
+            issues: const [],
+            designIssue: const [],
+          ),
         ];
 
         final report =
@@ -153,38 +212,24 @@ void main() {
         expect(report, isEmpty);
       });
 
-      test('with a lot of arguments', () {
+      test('without arguments', () {
         final records = [
           FileRecord(
-              fullPath: '/home/developer/work/project/example.dart',
-              relativePath: 'example.dart',
-              components: Map.unmodifiable(<String, ComponentRecord>{}),
-              functions: Map.unmodifiable(<String, FunctionRecord>{
-                'function': buildFunctionRecordStub(argumentsCount: 10),
-              }),
-              issues: const []),
+            fullPath: fullPath,
+            relativePath: 'example.dart',
+            components: Map.unmodifiable(<String, ComponentRecord>{}),
+            functions: Map.unmodifiable(<String, FunctionRecord>{
+              'function': buildFunctionRecordStub(argumentsCount: 0),
+            }),
+            issues: const [],
+            designIssue: const [],
+          ),
         ];
 
         final report =
-            (json.decode(_reporter.report(records).first) as List<Object>).first
-                as Map<String, Object>;
+            json.decode(_reporter.report(records).first) as List<Object>;
 
-        expect(report, containsPair('type', 'issue'));
-        expect(report, containsPair('check_name', 'numberOfArguments'));
-        expect(
-            report,
-            containsPair('description',
-                'Function `function` has 10 number of arguments (exceeds 4 allowed). Consider refactoring.'));
-        expect(report, containsPair('categories', ['Complexity']));
-        expect(
-            report,
-            containsPair('location', {
-              'path': 'example.dart',
-              'lines': {'begin': 0, 'end': 0},
-            }));
-        expect(report, containsPair('remediation_points', 50000));
-        expect(report,
-            containsPair('fingerprint', '9306995977c1febd3b6199617fbe68af'));
+        expect(report, isEmpty);
       });
     });
   });