+++ test/reporters/json_reporter_test.dart
@@ -5,6 +5,7 @@ import 'package:dart_code_metrics/src/models/code_issue.dart';
 import 'package:dart_code_metrics/src/models/code_issue_severity.dart';
 import 'package:dart_code_metrics/src/models/component_record.dart';
 import 'package:dart_code_metrics/src/models/config.dart';
+import 'package:dart_code_metrics/src/models/design_issue.dart';
 import 'package:dart_code_metrics/src/models/file_record.dart';
 import 'package:dart_code_metrics/src/models/function_record.dart';
 import 'package:dart_code_metrics/src/reporters/json_reporter.dart';
@@ -15,6 +16,8 @@ import '../stubs_builders.dart';
 
 void main() {
   group('JsonReporter.report report about', () {
+    const fullPath = '/home/developer/work/project/example.darts';
+
     JsonReporter _reporter;
 
     setUp(() {
@@ -29,7 +32,7 @@ void main() {
       test('aggregated arguments metric values', () {
         final records = [
           FileRecord(
-            fullPath: '/home/developer/work/project/example.dart',
+            fullPath: fullPath,
             relativePath: 'example.dart',
             components: Map.unmodifiable(<String, ComponentRecord>{
               'class': buildComponentRecordStub(methodsCount: 25),
@@ -42,6 +45,7 @@ void main() {
               'function3': buildFunctionRecordStub(argumentsCount: 10),
             }),
             issues: const [],
+            designIssue: const [],
           ),
         ];
 
@@ -53,6 +57,60 @@ void main() {
         expect(report, containsPair('total-number-of-arguments-violations', 2));
         expect(report, containsPair('average-number-of-methods', 13));
         expect(report, containsPair('total-number-of-methods-violations', 2));
+      });
+
+      test('with design issues', () {
+        const _issuePatternId = 'patternId1';
+        const _issuePatternDocumentation = 'https://docu.edu/patternId1.html';
+        const _issueLine = 2;
+        const _issueColumn = 3;
+        const _issueProblemCode = 'issue';
+        const _issueMessage = 'first issue message';
+        const _issueRecomendation = 'issue recomendation';
+
+        final records = [
+          FileRecord(
+            fullPath: fullPath,
+            relativePath: 'example.dart',
+            components: Map.unmodifiable(<String, ComponentRecord>{}),
+            functions: Map.unmodifiable(<String, FunctionRecord>{}),
+            issues: const [],
+            designIssue: [
+              DesignIssue(
+                patternId: _issuePatternId,
+                patternDocumentation: Uri.parse(_issuePatternDocumentation),
+                sourceSpan: SourceSpanBase(
+                    SourceLocation(1,
+                        sourceUrl: Uri.parse(fullPath),
+                        line: _issueLine,
+                        column: _issueColumn),
+                    SourceLocation(6, sourceUrl: Uri.parse(fullPath)),
+                    _issueProblemCode),
+                message: _issueMessage,
+                recommendation: _issueRecomendation,
+              ),
+            ],
+          ),
+        ];
+
+        final report =
+            (json.decode(_reporter.report(records).first) as List<Object>).first
+                as Map<String, Object>;
+
+        expect(report.containsKey('designIssues'), isTrue);
+
+        final issue = (report['designIssues'] as List<Object>)
+            .cast<Map<String, Object>>()
+            .single;
+
+        expect(issue, containsPair('patternId', _issuePatternId));
+        expect(issue,
+            containsPair('patternDocumentation', _issuePatternDocumentation));
+        expect(issue, containsPair('lineNumber', _issueLine));
+        expect(issue, containsPair('columnNumber', _issueColumn));
+        expect(issue, containsPair('problemCode', _issueProblemCode));
+        expect(issue, containsPair('message', _issueMessage));
+        expect(issue, containsPair('recommendation', _issueRecomendation));
       });
 
       test('with style severity issues', () {
@@ -67,7 +125,7 @@ void main() {
 
         final records = [
           FileRecord(
-            fullPath: '/home/developer/work/project/example.dart',
+            fullPath: fullPath,
             relativePath: 'example.dart',
             components: Map.unmodifiable(<String, ComponentRecord>{}),
             functions: Map.unmodifiable(<String, FunctionRecord>{}),
@@ -78,19 +136,17 @@ void main() {
                 severity: CodeIssueSeverity.style,
                 sourceSpan: SourceSpanBase(
                     SourceLocation(1,
-                        sourceUrl: Uri.parse(
-                            '/home/developer/work/project/example.dart'),
+                        sourceUrl: Uri.parse(fullPath),
                         line: _issueLine,
                         column: _issueColumn),
-                    SourceLocation(6,
-                        sourceUrl: Uri.parse(
-                            '/home/developer/work/project/example.dart')),
+                    SourceLocation(6, sourceUrl: Uri.parse(fullPath)),
                     _issueProblemCode),
                 message: _issueMessage,
                 correction: _issueCorrection,
                 correctionComment: _issueCorrectionComment,
               ),
             ],
+            designIssue: const [],
           ),
         ];
 
@@ -100,20 +156,21 @@ void main() {
 
         expect(report.containsKey('issues'), isTrue);
 
-        final issues =
-            (report['issues'] as List<Object>).cast<Map<String, Object>>();
+        final issue = (report['issues'] as List<Object>)
+            .cast<Map<String, Object>>()
+            .single;
 
-        expect(issues.single, containsPair('severity', 'style'));
-        expect(issues.single, containsPair('ruleId', _issueRuleId));
-        expect(issues.single,
-            containsPair('ruleDocumentation', _issueRuleDocumentation));
-        expect(issues.single, containsPair('lineNumber', _issueLine));
-        expect(issues.single, containsPair('columnNumber', _issueColumn));
-        expect(issues.single, containsPair('problemCode', _issueProblemCode));
-        expect(issues.single, containsPair('message', _issueMessage));
-        expect(issues.single, containsPair('correction', _issueCorrection));
-        expect(issues.single,
-            containsPair('correctionComment', _issueCorrectionComment));
+        expect(issue, containsPair('severity', 'style'));
+        expect(issue, containsPair('ruleId', _issueRuleId));
+        expect(
+            issue, containsPair('ruleDocumentation', _issueRuleDocumentation));
+        expect(issue, containsPair('lineNumber', _issueLine));
+        expect(issue, containsPair('columnNumber', _issueColumn));
+        expect(issue, containsPair('problemCode', _issueProblemCode));
+        expect(issue, containsPair('message', _issueMessage));
+        expect(issue, containsPair('correction', _issueCorrection));
+        expect(
+            issue, containsPair('correctionComment', _issueCorrectionComment));
       });
     });
 
@@ -121,13 +178,14 @@ void main() {
       test('without methods', () {
         final records = [
           FileRecord(
-            fullPath: '/home/developer/work/project/example.dart',
+            fullPath: fullPath,
             relativePath: 'example.dart',
             components: Map.unmodifiable(<String, ComponentRecord>{
               'class': buildComponentRecordStub(methodsCount: 0),
             }),
             functions: Map.unmodifiable(<String, FunctionRecord>{}),
             issues: const [],
+            designIssue: const [],
           ),
         ];
 
@@ -145,13 +203,14 @@ void main() {
       test('with a lot of methods', () {
         final records = [
           FileRecord(
-            fullPath: '/home/developer/work/project/example.dart',
+            fullPath: fullPath,
             relativePath: 'example.dart',
             components: Map.unmodifiable(<String, ComponentRecord>{
               'class': buildComponentRecordStub(methodsCount: 20),
             }),
             functions: Map.unmodifiable(<String, FunctionRecord>{}),
             issues: const [],
+            designIssue: const [],
           ),
         ];
 
@@ -168,16 +227,69 @@ void main() {
     });
 
     group('function', () {
+      test('with long body', () {
+        final records = [
+          FileRecord(
+            fullPath: fullPath,
+            relativePath: 'example.dart',
+            components: Map.unmodifiable(<String, ComponentRecord>{}),
+            functions: Map.unmodifiable(<String, FunctionRecord>{
+              'function': buildFunctionRecordStub(
+                  linesWithCode: List.generate(150, (index) => index)),
+            }),
+            issues: const [],
+            designIssue: const [],
+          ),
+        ];
+
+        final report =
+            (json.decode(_reporter.report(records).first) as List<Object>).first
+                as Map<String, Object>;
+        final functionReport = (report['records']
+            as Map<String, Object>)['function'] as Map<String, Object>;
+
+        expect(functionReport, containsPair('lines-of-executable-code', 150));
+        expect(functionReport,
+            containsPair('lines-of-executable-code-violation-level', 'alarm'));
+      });
+
+      test('with short body', () {
+        final records = [
+          FileRecord(
+            fullPath: fullPath,
+            relativePath: 'example.dart',
+            components: Map.unmodifiable(<String, ComponentRecord>{}),
+            functions: Map.unmodifiable(<String, FunctionRecord>{
+              'function': buildFunctionRecordStub(
+                  linesWithCode: List.generate(5, (index) => index)),
+            }),
+            issues: const [],
+            designIssue: const [],
+          ),
+        ];
+
+        final report =
+            (json.decode(_reporter.report(records).first) as List<Object>).first
+                as Map<String, Object>;
+        final functionReport = (report['records']
+            as Map<String, Object>)['function'] as Map<String, Object>;
+
+        expect(functionReport, containsPair('lines-of-executable-code', 5));
+        expect(functionReport,
+            containsPair('lines-of-executable-code-violation-level', 'none'));
+      });
+
       test('without arguments', () {
         final records = [
           FileRecord(
-            fullPath: '/home/developer/work/project/example.dart',
+            fullPath: fullPath,
             relativePath: 'example.dart',
             components: Map.unmodifiable(<String, ComponentRecord>{}),
             functions: Map.unmodifiable(<String, FunctionRecord>{
               'function': buildFunctionRecordStub(argumentsCount: 0),
             }),
             issues: const [],
+            designIssue: const [],
           ),
         ];
 
@@ -195,13 +307,14 @@ void main() {
       test('with a lot of arguments', () {
         final records = [
           FileRecord(
-            fullPath: '/home/developer/work/project/example.dart',
+            fullPath: fullPath,
             relativePath: 'example.dart',
             components: Map.unmodifiable(<String, ComponentRecord>{}),
             functions: Map.unmodifiable(<String, FunctionRecord>{
               'function': buildFunctionRecordStub(argumentsCount: 10),
             }),
             issues: const [],
+            designIssue: const [],
           ),
         ];